refactor(#302): spezza PipelineSidebarSections.tsx in moduli per dominio
File di 1098 righe che bundlava 4 componenti (Run/Pipelines/Document/
ExportDialogHost) più un pannello costi, ora un file per sezione sotto
PipelineSidebarSections/ con barrel index.ts per import invariati nei
consumer (PipelineSidebar.tsx, ProjectRailNext.tsx). Nessun cambio di
comportamento: tsc pulito, suite completa verde (570/570).
