+++ src/components/layout/PipelineSidebarSections/PipelineSidebarShell.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react';
+
+export function SidebarSectionShell({ children }: { children: ReactNode }) {
+  return <div className="px-1">{children}</div>;
+}
