+++ src/components/layout/PipelineSidebarSections/index.ts
@@ -1,0 +1,4 @@
+export { PipelineSidebarRunSection } from './PipelineSidebarRunSection';
+export { PipelineSidebarPipelinesSection } from './PipelineSidebarPipelinesSection';
+export { PipelineSidebarDocumentSection } from './PipelineSidebarDocumentSection';
+export { PipelineSidebarExportDialogHost } from './PipelineSidebarExportDialogHost';
